Move factories to app based modules
